Move telemetry send out of task into shutdown initializer + Update actionhero version
