+++ plugins/@grouparoo/app-templates/package.json
@@ -33,7 +33,7 @@
     "@types/jest": "*",
     "@types/node": "*",
     "@types/uuid": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "nock": "13.0.11",
     "prettier": "2.2.1",
