+++ plugins/@grouparoo/dbt/package.json
@@ -37,7 +37,7 @@
     "@grouparoo/spec-helper": "0.3.0-alpha.3",
     "@types/jest": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "nock": "13.0.11",
     "prettier": "2.2.1",
