+++ plugins/@grouparoo/demo/package.json
@@ -44,7 +44,7 @@
     "@grouparoo/spec-helper": "0.3.0-alpha.3",
     "@types/jest": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "ts-jest": "26.5.4",
     "typescript": "4.2.4"
