+++ plugins/@grouparoo/files-local/package.json
@@ -34,7 +34,7 @@
     "@grouparoo/core": "0.3.0-alpha.3",
     "@types/jest": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "prettier": "2.2.1",
     "ts-jest": "26.5.4",
