+++ plugins/@grouparoo/mysql/package.json
@@ -37,7 +37,7 @@
     "@types/jest": "*",
     "@types/mysql": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "csv-parse": "4.15.4",
     "jest": "26.6.3",
     "prettier": "2.2.1",
