+++ plugins/@grouparoo/postgres/package.json
@@ -40,7 +40,7 @@
     "@types/jest": "*",
     "@types/node": "*",
     "@types/pg": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "csv-parse": "4.15.4",
     "jest": "26.6.3",
     "prettier": "2.2.1",
