+++ plugins/@grouparoo/redshift/package.json
@@ -37,7 +37,7 @@
     "@types/jest": "*",
     "@types/node": "*",
     "@types/pg": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "prettier": "2.2.1",
     "ts-jest": "26.5.4",
