+++ plugins/@grouparoo/sentry/package.json
@@ -36,7 +36,7 @@
     "@grouparoo/spec-helper": "0.3.0-alpha.3",
     "@types/jest": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "jest": "26.6.3",
     "prettier": "2.2.1",
     "ts-jest": "26.5.4",
