+++ plugins/@grouparoo/spec-helper/package.json
@@ -40,7 +40,7 @@
     "@types/faker": "*",
     "@types/jest": "*",
     "@types/node": "*",
-    "actionhero": "25.0.12",
+    "actionhero": "25.0.14",
     "fs-extra": "9.1.0",
     "jest": "26.6.3",
     "prettier": "2.2.1",
